inline docs for responsive-container; readme template consumable by react-docgen; readme
