+++ src/ui/responsive-container/_readme.template.md
@@ -1,0 +1,6 @@
+<ResponsiveContainer \/>
+=====================
+
+{{{responsive-container.description}}}
+
+{{{responsive-container.props}}}
