+++ src/ui/responsive-container/index.js
@@ -1,1 +1,1 @@
-export { default as default } from './src/index';
+export { default as default } from './src/responsive-container';
